adding tests that require manual setup
i.e. to verify encoding issues are not caused by this library
